resolves list-to-products call
