+++ client/src/components/ListCard.js
@@ -51,26 +51,46 @@ const ListCard = (props) => {
   const classes = useStyles();
   const [listOpen, setListOpen] = useState(false);
   const [itemCount, setItemCount] = useState('');
+  const [listToProducts, setListToProducts] = useState([]); // list-to-products table {list_id, product_id}
+  const [products, setProducts] = useState([]); // products table {product_id, img_url, price}
+  const [addProductOpen, setAddProductOpen] = useState(false);
 
+  // This affects the Modal open component which opens the modal when set to True
   const changeListOpen = () => {
     setListOpen(!listOpen);
-  };
-  const [addProductOpen, setAddProductOpen] = useState(false);
-  const changeAddProductOpen = () => {
-    setAddProductOpen((previous) => !previous);
+    getProductIds();
   };
 
-  const getItemCount = async () => {
+  // Gets the total number of items in the list
+  const getProductIds = async () => {
     const res = await fetch(`/list-to-products/${listId}`)
     const json = await res.json();
     setItemCount(json.length)
+    setListToProducts(json);
+  };
+
+  const getProducts = async () => {
+    const newProducts = await Promise.all(
+      listToProducts.map(async (relation) => {
+        const res = await fetch(`/products/${relation.product_id}`);
+        return res.json(res);
+      })
+    );
+    setProducts(newProducts);
   };
 
   useEffect(() => {
-    getItemCount()
-  }, [listOpen,listChange]);
+    getProducts();
+  }, [listToProducts]);
+
+  useEffect(() => {
+    getProductIds()
+  }, [listChange]);
 
 
+  // const changeAddProductOpen = () => {
+  //   setAddProductOpen((previous) => !previous);
+  // };
   return (
     <Box>
       <Box onClick={changeListOpen} className={classes.listContainer}>
@@ -83,7 +103,7 @@ const ListCard = (props) => {
           >{`${itemCount} items`}</Typography>
         </Box>
       </Box>
-      <ListPopup {...{listId, listTitle, itemCount, listOpen, changeListOpen }} />
+      <ListPopup {...{products, listId, getProductIds, listTitle, itemCount, listOpen, changeListOpen }} />
       {/* <AddProduct
         {...{
           listTitle,
